requirements.txt
batchgenerators==0.19.3   
certifi==2019.6.16
cffi==1.12.3   
cycler==0.10.0   
decorator==4.4.0    
future==0.17.1   
h5py==2.9.0    
imageio==2.5.0    
joblib==0.13.2   
kiwisolver==1.1.0    
linecache2==1.0.0    
matplotlib==3.1.1    
networkx==2.3      
nibabel==2.4.1    
nilearn==0.5.0    
numpy==1.16.4   
pandas==0.25.0   
Pillow==6.1.0    
pip==19.1.1   
pycparser==2.19     
pyparsing==2.4.0    
python-dateutil==2.8.0    
pytz==2019.1   
PyWavelets==1.0.3    
scikit-image==0.15.0   
scikit-learn==0.21.2   
scipy==1.3.0    
setuptools==41.0.1   
six==1.12.0   
telegram==0.0.1    
torch==1.0.1    
traceback2==1.4.0    
unittest2==1.1.0    
wheel==0.33.4
pyprg==0.1
pyyaml
visdom
